Create parent for Jackson-based configurations
