+++ pom.xml
@@ -236,6 +236,7 @@
     <module>log4j-api</module>
     <module>log4j-api-test</module>
     <module>log4j-appserver</module>
+    <module>log4j-config-jackson</module>
     <module>log4j-config-yaml</module>
     <module>log4j-core</module>
     <module>log4j-core-its</module>
@@ -372,6 +373,12 @@
       <dependency>
         <groupId>org.apache.logging.log4j</groupId>
         <artifactId>log4j-appserver</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+
+      <dependency>
+        <groupId>org.apache.logging.log4j</groupId>
+        <artifactId>log4j-config-jackson</artifactId>
         <version>${project.version}</version>
       </dependency>
